% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureImportance.R
\name{compute_integratedVariances}
\alias{compute_integratedVariances}
\title{Compute Integrated Variances}
\usage{
compute_integratedVariances(obj, exps, X, n, p, K, nSamp, nMC)
}
\description{
Compute Integrated Variances
}
\keyword{internal}
