% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MVmix.R
\name{est_lag}
\alias{est_lag}
\title{Estimate lagged effects}
\usage{
est_lag(obj, Xhold = NULL, Xshift = 1, fixomega = FALSE, fixomegaval = NULL)
}
\arguments{
\item{obj}{fitted model}

\item{Xhold}{value to compare to and for other exposures to be set to. NULL defaults to median}

\item{Xshift}{increase in X for contrasts}

\item{fixomega}{hold weights fixed (default FALSE)}

\item{fixomegaval}{fixed weight values (if fixomega=TRUE)}
}
\value{
lagged effect estimates
}
\description{
Estimate effect of 1 unit change in x_lag for each different lag holding others constant
}
