% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExposureImportance.R
\name{ExposureImportance}
\alias{ExposureImportance}
\title{Construct MCMC sampler}
\usage{
ExposureImportance(obj, exposures, nMC = 250, nSamp = NULL)
}
\arguments{
\item{obj}{fitted model}

\item{exposures}{which exposures to report}

\item{nMC}{No. of MC samples}

\item{nSamp}{No. of exposure samples}
}
\value{
function for MCMC update
}
\description{
Internal function to construct the MCMC sampling function
}
