% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.saddle.R
\name{fb.saddle}
\alias{fb.saddle}
\title{Saddle point approximation of FB distribution.}
\usage{
fb.saddle(gam, lam)
}
\description{
By: Tsagris Michail 02/2014.
mtsagris@yahoo.gr.
References: Kume Alfred and Wood Andrew T.A. (2005)
Saddlepoint approximations for the Bingham and Fisher-Bingham normalizing constants (Biometrika)
}
\keyword{internal}
