% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_B_beta}
\alias{get_B_beta}
\title{Compute B times the corresponding beta for all obs (d)}
\usage{
get_B_beta(params, const)
}
\description{
Compute B times the corresponding beta for all obs (d)
}
\keyword{internal}
