% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_B_beta_k}
\alias{get_B_beta_k}
\title{Compute B times the corresponding beta for single cluster}
\usage{
get_B_beta_k(params, const, kk)
}
\description{
Compute B times the corresponding beta for single cluster
}
\keyword{internal}
