% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_B_beta_k_b}
\alias{get_B_beta_k_b}
\title{Get Btheta times beta (alternate version for updating Ztheta=b in update_clustMemb)}
\usage{
get_B_beta_k_b(params, const, kk, Ztheta)
}
\description{
Get Btheta times beta (alternate version for updating Ztheta=b in update_clustMemb)
}
\keyword{internal}
