% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_Btheta_b}
\alias{get_Btheta_b}
\title{Get Btheta (alternate version for updating Ztheta=b in update_clustMemb)}
\usage{
get_Btheta_b(Xomega, const, Ztheta = NULL, k, j)
}
\description{
Get Btheta (alternate version for updating Ztheta=b in update_clustMemb)
}
\keyword{internal}
