% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_DerivB_beta}
\alias{get_DerivB_beta}
\title{Compute DerivB times the corresponding beta for given k,j}
\usage{
get_DerivB_beta(params, const, kk, jj)
}
\description{
Compute DerivB times the corresponding beta for given k,j
}
\keyword{internal}
