% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_DerivBtheta}
\alias{get_DerivBtheta}
\title{Get derivatives of Basis functions}
\usage{
get_DerivBtheta(Xomega, const, params, k, j)
}
\description{
Get derivatives of Basis functions
}
\keyword{internal}
