% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_Vlogdensity}
\alias{get_Vlogdensity}
\title{Computes density of V_c ## up to proportionality constant}
\usage{
get_Vlogdensity(vv, cc, params, Vbeta = TRUE, const)
}
\description{
Computes density of V_c ## up to proportionality constant
}
\keyword{internal}
