% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_XTyhat}
\alias{get_XTyhat}
\title{Get components for omegastar sampling}
\usage{
get_XTyhat(cc, whichk, whichkj, params, const)
}
\description{
Get components for omegastar sampling
}
\keyword{internal}
