% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initialize_const}
\alias{initialize_const}
\title{Initialize constants}
\usage{
initialize_const(
  Y,
  X,
  Z,
  niter,
  nburn,
  nthin,
  nchains,
  ncores,
  clustering,
  fixedZbeta,
  fixedZtheta,
  maxClusters,
  prior_alpha_beta,
  prior_alpha_theta,
  prior_rho,
  prior_tau_theta,
  prior_lambda_beta,
  prior_lambda_theta,
  prior_xi,
  prior_sigma2,
  prop_phi_a,
  sharedlambda,
  DLM,
  DLMpenalty,
  lagOrder,
  diff,
  MIM,
  MIMorder,
  LM,
  betaOrder,
  stepsize_logrho,
  stepsize_loglambda_theta,
  stepsize_logxi,
  stepsize_logsigma2,
  Vgridsearch,
  gridsize,
  rfbtries,
  approx
)
}
\description{
Initialize constants
}
\keyword{internal}
