% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_clusters.R
\name{make_heatplot}
\alias{make_heatplot}
\title{Summarize pairwise clustering}
\usage{
make_heatplot(heat)
}
\arguments{
\item{heat}{heat matrix (matrix of pairwise clustering probabilities) generated by 'pairwise_clusters'}
}
\value{
heatplot of pairwise clustering probabilities
}
\description{
Function to summarize pairwise clustering
}
