% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVmix.R
\name{MVmix}
\alias{MVmix}
\title{Fit multivariate mixtures model with adaptive clustering}
\usage{
MVmix(
  Y,
  X,
  Z,
  niter = 10000,
  nburn = 0.5 * niter,
  nthin = 10,
  nchains = 1,
  ncores = 1,
  clustering = "both",
  fixedZbeta = NULL,
  fixedZtheta = NULL,
  maxClusters = NULL,
  prior_alpha_beta = c(1, 1),
  prior_alpha_theta = c(1, 1),
  prior_rho = c(1, 1),
  prior_tau_theta = 0,
  prior_lambda_beta = c(1, 1),
  prior_lambda_theta = c(1, 0.001),
  prior_xi = c(0.01, 0.01),
  prior_sigma2 = c(0.01, 0.01),
  prop_phi_a = 200,
  sharedlambda = TRUE,
  DLM = FALSE,
  DLMpenalty = FALSE,
  lagOrder = 6,
  diff = 2,
  MIM = FALSE,
  MIMorder = 4,
  LM = FALSE,
  betaOrder = -1,
  stepsize_logrho = 1,
  stepsize_loglambda_theta = 1,
  stepsize_logxi = 1,
  stepsize_logsigma2 = 1,
  Vgridsearch = TRUE,
  gridsize = 10,
  rfbtries = 1000,
  approx = TRUE,
  appendchain = NULL
)
}
\arguments{
\item{Y}{n x K matrix of responses}

\item{X}{p-list of n x L matrix of exposures}

\item{Z}{confounders to be adjusted}

\item{niter}{number of iterations}

\item{nburn}{burn-in fraction}

\item{nthin}{thinning number}

\item{nchains}{number of chains ## not yet implemented}

\item{ncores}{number of cores for mclapply (set to 1 for non-parallel) ## only used if nchains>1}

\item{clustering}{one of "both","theta","beta" or ,"neither";}

\item{fixedZbeta}{optional KxP matrix of fixed cluster IDs for beta; only used if clustering="theta" or "neither"}

\item{fixedZtheta}{optional KxP matrix of fixed cluster IDs for theta; only used if clustering="theta" or "neither"}

\item{maxClusters}{max number of distinct clusters}

\item{prior_alpha_beta}{gamma prior hyperparameters for alpha_beta}

\item{prior_alpha_theta}{gamma prior hyperparameters for alpha_theta}

\item{prior_rho}{gamma prior hyperparameters for rho}

\item{prior_tau_theta}{Do not change. prior hyperparameter for thetastar direction}

\item{prior_lambda_beta}{gamma prior hyperparameters for lambda_beta controlling smoothness of exposure response functions}

\item{prior_lambda_theta}{gamma prior hyperparameters for lambda_theta controlling smoothness of weight functions (if DLMpenalty=TRUE)}

\item{prior_xi}{inverse gamma prior hyperparameters for xi}

\item{prior_sigma2}{inverse gamma prior hyperparameters for sigma^2}

\item{prop_phi_a}{hyperparameter a for the beta(a,b) proposal on phistar. Higher value means smaller steps.}

\item{sharedlambda}{Should a single lambda value controlling smoothness be shared across all exposure-response functions? T/F}

\item{DLM}{Use B-spline approximation to impose smoothness in weights over time}

\item{DLMpenalty}{Include smoothness penalty in weights over time; only if DLM=TRUE}

\item{lagOrder}{Number of basis functions for weights (if DLM=TRUE); NULL indicates no dimension reduction}

\item{diff}{Degree of difference penalty matrix (if DLMpenalty=TRUE)}

\item{MIM}{Fit MIM version with unknown index structure (T/F)}

\item{MIMorder}{Maximum order of MIM (i.e. number of indices); ignored if MIM=FALSE}

\item{LM}{Force linear exposure response relationships (T/F)}

\item{betaOrder}{B-spline basis dimension for exposure response functions. default -1 allows mgcv to choose automatically}

\item{stepsize_logrho}{SD for random walk updates on log(rho)}

\item{stepsize_loglambda_theta}{SD for random walk updates on log(lambda_theta)}

\item{stepsize_logxi}{SD for random walk updates on log(xi)}

\item{stepsize_logsigma2}{SD for random walk updates on log(sigma2)}

\item{Vgridsearch}{Use grid search for approximate sampling of V_c}

\item{gridsize}{Size of grid. Not used if Vgridsearch=FALSE}

\item{rfbtries}{Max number of tries for drawing from Fisher-Bingham; equiv.to "mtop" for rFisherBingham (default 1000)}

\item{approx}{Should rFB approximation be used? Otherwise use polar transformation and MH updates}

\item{appendchain}{Option to use last values of a previous fit as starting values for a new chain.}
}
\value{
Posterior samples for all parameters.
}
\description{
This function will take in the observed data and fit a possibly multivariate
additive mixtures model with or without adaptive clustering over outcomes/exposures.
}
