% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_clusters.R
\name{pairwise_clusters}
\alias{pairwise_clusters}
\title{Summarize pairwise clustering}
\usage{
pairwise_clusters(obj)
}
\arguments{
\item{obj}{fitted model object}
}
\value{
list of pairwise clustering probability matrices for beta and theta.
}
\description{
Function to summarize pairwise clustering
}
