% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MVmix.R
\name{predict_MVmix}
\alias{predict_MVmix}
\title{Predict/estimate exposure response surface}
\usage{
predict_MVmix(
  obj,
  newX = NULL,
  newZ = NULL,
  fixomega = FALSE,
  fixomegaval = NULL,
  include_intercept = TRUE,
  allx = FALSE,
  contrast = FALSE
)
}
\arguments{
\item{obj}{fitted model object}

\item{newX}{optional list of new exposure values}

\item{newZ}{optional matrix of new covariates; if null, dont include covariate effects}

\item{fixomega}{hold weights fixed (default FALSE)}

\item{fixomegaval}{fixed weight values (if fixomega=TRUE)}

\item{include_intercept}{include intercept in predictions (T/F)}

\item{allx}{combine all x simultaneously or do individual exposure prediction (T/F)}

\item{contrast}{report contrasts (T/F)}
}
\value{
list of predictions/fitted values
}
\description{
Function for estimating fitted/predicted values
}
