% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MVmix.R
\name{summarize_pred_all}
\alias{summarize_pred_all}
\title{Summarize all predictions with intercept handling}
\usage{
summarize_pred_all(pred, obj, include_intercept, newZ)
}
\description{
Summarize all predictions with intercept handling
}
\keyword{internal}
