% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_loglambda_theta}
\alias{update_loglambda_theta}
\title{Update smoothness penalty on theta}
\usage{
update_loglambda_theta(params, const)
}
\description{
Update smoothness penalty on theta
}
\keyword{internal}
