% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_thetastar_MH_beta}
\alias{update_thetastar_MH_beta}
\title{Update thetastar via MH step on polar coordinates}
\usage{
update_thetastar_MH_beta(params, const)
}
\description{
Update thetastar via MH step on polar coordinates
}
\keyword{internal}
